+++ YasinDavids_ST10269809_PROG6221_POE_PART1/Menu.cs
@@ -47,7 +47,7 @@ namespace YasinDavids_ST10269809_PROG6221_POE_PART1
                             Console.WriteLine("(0.5)\n(1)\n(2)\n(3)");
                             // Takes input to scale
 
-                            int scale = int.Parse(Console.ReadLine());
+                            double scale = double.Parse(Console.ReadLine());
                             //Scales the ingredients according to user choice
                             if (scale == 0.5 || scale == 1 || scale == 2 || scale == 3)
                             {
